requirements.txt
Flask
pandas
geopandas
folium
google-cloud-bigquery
deep-translator
